requirements.txt
compreface-sdk>=0.6.0